Get tests building on netfx
