+++ src/libraries/System.Utf8String.Experimental/tests/System/MemoryTests.netcoreapp.cs
@@ -1,0 +1,72 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Buffers;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using Xunit;
+
+using static System.Tests.Utf8TestUtilities;
+
+namespace System.Tests
+{
+    public partial class MemoryTests
+    {
+        [Fact]
+        public static void MemoryMarshal_TryGetArrayOfChar8_Utf8String()
+        {
+            ReadOnlyMemory<Char8> rom = u8("Hello").AsMemory();
+
+            Assert.False(MemoryMarshal.TryGetArray(rom, out ArraySegment<Char8> segment));
+            Assert.True(default(ArraySegment<Char8>).Equals(segment));
+        }
+
+        [Fact]
+        public static unsafe void MemoryOfChar8_WithUtf8String_Pin()
+        {
+            Utf8String theString = u8("Hello");
+            ReadOnlyMemory<Char8> rom = theString.AsMemory();
+            MemoryHandle memHandle = default;
+            try
+            {
+                memHandle = Unsafe.As<ReadOnlyMemory<Char8>, Memory<Char8>>(ref rom).Pin();
+                Assert.True(memHandle.Pointer == Unsafe.AsPointer(ref Unsafe.AsRef(in theString.GetPinnableReference())));
+            }
+            finally
+            {
+                memHandle.Dispose();
+            }
+        }
+
+        [Fact]
+        public static void MemoryOfChar8_WithUtf8String_ToString()
+        {
+            ReadOnlyMemory<Char8> rom = u8("Hello").AsMemory();
+            Assert.Equal("Hello", Unsafe.As<ReadOnlyMemory<Char8>, Memory<Char8>>(ref rom).ToString());
+        }
+
+        [Fact]
+        public static unsafe void ReadOnlyMemoryOfChar8_WithUtf8String_Pin()
+        {
+            Utf8String theString = u8("Hello");
+            ReadOnlyMemory<Char8> rom = theString.AsMemory();
+            MemoryHandle memHandle = default;
+            try
+            {
+                memHandle = rom.Pin();
+                Assert.True(memHandle.Pointer == Unsafe.AsPointer(ref Unsafe.AsRef(in theString.GetPinnableReference())));
+            }
+            finally
+            {
+                memHandle.Dispose();
+            }
+        }
+
+        [Fact]
+        public static void ReadOnlyMemoryOfChar8_WithUtf8String_ToString()
+        {
+            Assert.Equal("Hello", u8("Hello").AsMemory().ToString());
+        }
+    }
+}
