+++ src/libraries/System.Utf8String.Experimental/tests/System/Utf8ExtensionsTests.cs
@@ -24,20 +24,12 @@ namespace System.Tests
             // Next, an empty but non-default span should become an empty but non-default span.
 
             Assert.True(new ReadOnlySpan<byte>((void*)0x12345, 0) == new ReadOnlySpan<Char8>((void*)0x12345, 0).AsBytes());
-
-            // Finally, a span wrapping data should become a span wrapping that same data.
-
-            Utf8String theString = u8("Hello");
-            Assert.True(MemoryMarshal.CreateReadOnlySpan(ref Unsafe.AsRef(in theString.GetPinnableReference()), 5) == (theString.AsMemory().Span).AsBytes());
         }
 
         [Fact]
         public void AsBytes_FromUtf8String()
         {
             Assert.True(default(ReadOnlySpan<byte>) == ((Utf8String)null).AsBytes());
-
-            Utf8String theString = u8("Hello");
-            Assert.True(MemoryMarshal.CreateReadOnlySpan(ref Unsafe.AsRef(in theString.GetPinnableReference()), 5) == theString.AsBytes());
         }
 
         [Fact]
@@ -76,56 +68,18 @@ namespace System.Tests
         }
 
         [Fact]
-        public void AsMemory_FromUtf8String()
-        {
-            Assert.True(default(ReadOnlyMemory<Char8>).Equals(((Utf8String)null).AsMemory()));
-
-            Utf8String theString = u8("Hello");
-            Assert.True(MemoryMarshal.CreateReadOnlySpan(ref Unsafe.As<byte, Char8>(ref Unsafe.AsRef(in theString.GetPinnableReference())), 5) == theString.AsMemory().Span);
-        }
-
-        [Fact]
-        public void AsMemory_FromUtf8String_WithStart()
-        {
-            Assert.True(default(ReadOnlyMemory<Char8>).Equals(((Utf8String)null).AsMemory(0)));
-            Assert.True(u8("Hello").AsMemory(5).IsEmpty);
-
-            SpanAssert.Equal(new Char8[] { (Char8)'e', (Char8)'l', (Char8)'l', (Char8)'o' }, u8("Hello").AsMemory(1).Span);
-        }
-
-        [Fact]
-        public void AsMemory_FromUtf8String_WithStart_ArgOutOfRange()
-        {
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => ((Utf8String)null).AsMemory(1));
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => u8("Hello").AsMemory(-1));
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => u8("Hello").AsMemory(6));
-        }
-
-        [Fact]
-        public void AsMemory_FromUtf8String_WithStartAndLength()
-        {
-            Assert.True(default(ReadOnlyMemory<Char8>).Equals(((Utf8String)null).AsMemory(0, 0)));
-            Assert.True(u8("Hello").AsMemory(5, 0).IsEmpty);
-
-            SpanAssert.Equal(new Char8[] { (Char8)'e', (Char8)'l', (Char8)'l' }, u8("Hello").AsMemory(1, 3).Span);
-        }
-
-        [Fact]
-        public void AsMemory_FromUtf8String_WithStartAndLength_ArgOutOfRange()
-        {
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => ((Utf8String)null).AsMemory(0, 1));
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => ((Utf8String)null).AsMemory(1, 0));
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => u8("Hello").AsMemory(5, 1));
-            Assert.Throws<ArgumentOutOfRangeException>("start", () => u8("Hello").AsMemory(4, -2));
-        }
-
-        [Fact]
-        public void AsMemoryBytes_FromUtf8String()
+        public unsafe void AsMemoryBytes_FromUtf8String()
         {
             Assert.True(default(ReadOnlyMemory<byte>).Equals(((Utf8String)null).AsMemoryBytes()));
 
             Utf8String theString = u8("Hello");
-            Assert.True(MemoryMarshal.CreateReadOnlySpan(ref Unsafe.AsRef(in theString.GetPinnableReference()), 5) == theString.AsMemoryBytes().Span);
+            fixed (byte* pTheString = theString)
+            {
+                fixed (byte* pTheStringAsMemoryBytes = theString.AsMemoryBytes().Span)
+                {
+                    Assert.True(pTheString == pTheStringAsMemoryBytes);
+                }
+            }
         }
 
         [Fact]
