+++ src/libraries/System.Utf8String.Experimental/tests/System/Utf8SpanTests.cs
@@ -287,7 +287,7 @@ namespace System.Text.Tests
             {
                 using BoundedMemory<char> boundedMemory = BoundedMemory.Allocate<char>(i);
                 Assert.Equal(expected.Length, span.ToChars(boundedMemory.Span));
-                Assert.True(boundedMemory.Span.Slice(0, expected.Length).SequenceEqual(expected));
+                Assert.True(boundedMemory.Span.Slice(0, expected.Length).SequenceEqual(expected.AsSpan()));
             }
         }
